#53 add more verbose information to rnpm script

+++ src/scripts/prelink.js
@@ -11,14 +11,21 @@ if (!hasNecessaryFile) {
 
 var package = fs.readFileSync(PACKAGE_JSON);
 var APP_NAME = package.name;
-var VERSION = parseFloat(/\d\.\d+(?=\.)/.exec(package.dependencies['react-native']));
 var APPLICATION_MAIN = process.cwd() + '/android/app/src/main/java/com/' + APP_NAME.toLocaleLowerCase() + '/MainApplication.java';
 
-if(VERSION >= 0.29) {
+if(!fs.existsSync('APPLICATION_MAIN')) {
+  throw 'RNFetchBlob could not found link Android automatically, MainApplication.java not found in path : ' + APPLICATION_MAIN
+}
 
+var VERSION = checkVersion();
+console.log('RNFetchBlob detected app version .. ' + VERSION);
+
+if(VERSION >= 0.29) {
+  console.log('RNFetchBlob patching MainApplication.java .. ');
   var main = fs.readFileSync(APPLICATION_MAIN);
   main = main.replace('new MainReactPackage()', 'new RNFetchBlobPackage(),\n           new MainReactPackage()');
   fs.writeFileSync(APPLICATION_MAIN, main);
+  console.log('RNFetchBlob patching MainApplication.java .. ok')
 
 }
 
@@ -28,6 +35,8 @@ fs.readFile(MANIFEST_PATH, function(err, data) {
   if(err)
     console.log('failed to locate AndroidManifest.xml file, you may have to add file access permission manually.');
   else {
+
+    console.log('RNFetchBlob patching AndroidManifest.xml .. ');
     // append fs permission
     data = String(data).replace(
       '<uses-permission android:name="android.permission.INTERNET" />',
@@ -39,7 +48,13 @@ fs.readFile(MANIFEST_PATH, function(err, data) {
       '<category android:name="android.intent.category.LAUNCHER" />\n     <action android:name="android.intent.action.DOWNLOAD_COMPLETE"/>'
     )
     fs.writeFileSync(MANIFEST_PATH, data);
+    console.log('RNFetchBlob patching AndroidManifest.xml .. ok');
 
   }
 
 })
+
+function checkVersion() {
+  console.log('RNFetchBlob checking app version ..');
+  return parseFloat(/\d\.\d+(?=\.)/.exec(package.dependencies['react-native']));
+}
